Added Variable command reporter without changing output
