+++ src/robotframework_find_unused/commands/step/variables_definitions.py
@@ -1,16 +1,8 @@
 from pathlib import Path
 
-import click
-
-from robotframework_find_unused.common.const import (
-    DONE_MARKER,
-    ERROR_MARKER,
-    INDENT,
-    VERBOSE_NO,
-    VERBOSE_SINGLE,
-    VariableData,
-)
+from robotframework_find_unused.common.const import VariableData
 from robotframework_find_unused.common.normalize import normalize_variable_name
+from robotframework_find_unused.reporter.base.variable_reporter import VariableReporter
 from robotframework_find_unused.resolve.resolve_variables import resolve_variable_name
 from robotframework_find_unused.visitors.robot import visit_robot_files
 from robotframework_find_unused.visitors.robot.variable_definition import (
@@ -22,26 +14,20 @@ def cli_get_variable_definitions(
     file_paths: list[Path],
     source_path: Path,
     *,
-    verbose: int,
+    reporter: VariableReporter,
 ):
     """
     Walk through all robot files to discover non-local variable definitions and show progress
     """
-    click.echo("Gathering variables definitions...")
-    variables = _get_variable_definitions(file_paths, source_path)
+    reporter.on_get_variable_definitions_start(file_paths, source_path)
 
-    _log_variable_stats(list(variables.values()), verbose)
-    return variables
-
-
-def _get_variable_definitions(file_paths: list[Path], source_path: Path) -> dict[str, VariableData]:
-    """
-    Walk through all robot files to discover non-local variable definitions.
-    """
     visitor = RobotVisitorVariableDefinitions(source_path, set(file_paths))
     visit_robot_files(file_paths, visitor)
 
-    return _resolve_vars_in_var_name(visitor.variables)
+    variables = _resolve_vars_in_var_name(visitor.variables)
+
+    reporter.on_get_variable_definitions_end(file_paths, source_path, variables)
+    return variables
 
 
 def _resolve_vars_in_var_name(variables: dict[str, VariableData]) -> dict[str, VariableData]:
@@ -76,34 +62,3 @@ def _resolve_vars_in_var_name(variables: dict[str, VariableData]) -> dict[str, V
         resolved_variables[used_var].use_count += 1
 
     return resolved_variables
-
-
-def _log_variable_stats(variables: list[VariableData], verbose: int) -> None:
-    """
-    Output details to the user
-    """
-    click.echo(
-        (ERROR_MARKER if len(variables) == 0 else DONE_MARKER)
-        + f" Found {len(variables)} unique non-local variables definitions",
-    )
-
-    if verbose == VERBOSE_NO:
-        return
-
-    var_types: dict[str, list[str]] = {}
-    for var in variables:
-        if var.defined_in_type not in var_types:
-            var_types[var.defined_in_type] = []
-        var_types[var.defined_in_type].append(var.name)
-
-    for defined_in_type, var_names in sorted(
-        var_types.items(),
-        key=lambda items: len(items[1]),
-        reverse=True,
-    ):
-        click.echo(f"{INDENT}{len(var_names)} variables definitions of type '{defined_in_type}'")
-
-        if verbose == VERBOSE_SINGLE:
-            continue
-        for name in var_names:
-            click.echo(f"{INDENT}{INDENT}{click.style(name, fg='bright_black')}")
