+++ src/robotframework_find_unused/commands/variables/__init__.py
@@ -1,0 +1,11 @@
+"""
+Find unused variables command
+"""
+
+from .options import VariableOptions
+from .variables import cli_variables
+
+__all__ = [
+    "VariableOptions",
+    "cli_variables",
+]
